remove mount icon on --kill-kbfs
