+++ go/libkb/util_windows.go
@@ -15,7 +15,10 @@ import (
 	"time"
 	"unsafe"
 
+	"github.com/kardianos/osext"
+
 	"golang.org/x/sys/windows"
+	"golang.org/x/sys/windows/registry"
 )
 
 type GUID struct {
@@ -40,6 +43,7 @@ var (
 	modOle32                 = windows.NewLazySystemDLL("Ole32.dll")
 	procSHGetKnownFolderPath = modShell32.NewProc("SHGetKnownFolderPath")
 	procCoTaskMemFree        = modOle32.NewProc("CoTaskMemFree")
+	shChangeNotifyProc       = modShell32.NewProc("SHChangeNotify")
 )
 
 // LookPath searches for an executable binary named file
@@ -183,4 +187,48 @@ func RemoteSettingsRepairman(g *GlobalContext) error {
 		}
 	}
 	return nil
+}
+
+// Notify the shell that the thing located at path has changed
+func notifyShell(path string) {
+	shChangeNotifyProc.Call(
+		uintptr(0x00002000), // SHCNE_UPDATEITEM
+		uintptr(0x0005),     // SHCNF_PATHW
+		uintptr(unsafe.Pointer(syscall.StringToUTF16Ptr(path))),
+		0)
+}
+
+// Manipulate registry entries to reflect the mount point icon in the shell
+func ChangeMountIcon(oldMount string, newMount string) error {
+	if oldMount != "" {
+		// DeleteKey doesn't work if there are subkeys
+		registry.DeleteKey(registry.CURRENT_USER, `SOFTWARE\Classes\Applications\Explorer.exe\Drives\`+oldMount[:1]+`\DefaultIcon`)
+		registry.DeleteKey(registry.CURRENT_USER, `SOFTWARE\Classes\Applications\Explorer.exe\Drives\`+oldMount[:1]+`\DefaultLabel`)
+		registry.DeleteKey(registry.CURRENT_USER, `SOFTWARE\Classes\Applications\Explorer.exe\Drives\`+oldMount[:1])
+		notifyShell(oldMount)
+	}
+	if newMount == "" {
+		return nil
+	}
+	k, _, err := registry.CreateKey(registry.CURRENT_USER, `SOFTWARE\Classes\Applications\Explorer.exe\Drives\`+newMount[:1]+`\DefaultIcon`, registry.SET_VALUE|registry.CREATE_SUB_KEY|registry.WRITE)
+	defer k.Close()
+	if err != nil {
+		return err
+	}
+	keybaseExe, err := osext.Executable()
+	if err != nil {
+		return err
+	}
+	// Use the second icon bound into keybase.exe - hence the 1
+	err = k.SetStringValue("", keybaseExe+",1")
+	if err != nil {
+		return err
+	}
+
+	// Also give a nice label
+	k2, _, err := registry.CreateKey(registry.CURRENT_USER, `SOFTWARE\Classes\Applications\Explorer.exe\Drives\`+newMount[:1]+`\DefaultLabel`, registry.SET_VALUE|registry.CREATE_SUB_KEY|registry.WRITE)
+	defer k2.Close()
+	err = k2.SetStringValue("", "Keybase")
+	notifyShell(newMount)
+	return err
 }
