+++ go/service/kbfs_mount_windows.go
@@ -7,20 +7,17 @@ package service
 
 import (
 	"fmt"
-	"github.com/kardianos/osext"
-	"golang.org/x/sys/windows"
-	"golang.org/x/sys/windows/registry"
 	"strings"
 	"syscall"
 	"unsafe"
+
+	"golang.org/x/sys/windows"
 )
 
 var (
 	kernel32DLL        = windows.NewLazySystemDLL("kernel32.dll")
 	getVolumeProc      = kernel32DLL.NewProc("GetVolumeInformationW")
 	queryDosDeviceProc = kernel32DLL.NewProc("QueryDosDeviceW")
-	shell32DLL         = windows.NewLazySystemDLL("shell32.dll")
-	shChangeNotifyProc = shell32DLL.NewProc("SHChangeNotify")
 )
 
 // getVolumeName requires a drive letter and colon with a
@@ -99,48 +96,4 @@ func getMountDirs() ([]string, error) {
 		err = fmt.Errorf("No drive letters available")
 	}
 	return drives, err
-}
-
-// Notify the shell that the thing located at path has changed
-func notifyShell(path string) {
-	shChangeNotifyProc.Call(
-		uintptr(0x00002000), // SHCNE_UPDATEITEM
-		uintptr(0x0005),     // SHCNF_PATHW
-		uintptr(unsafe.Pointer(syscall.StringToUTF16Ptr(path))),
-		0)
-}
-
-// Manipulate registry entries to reflect the mount point icon in the shell
-func doMountChange(oldMount string, newMount string) error {
-	if oldMount != "" {
-		// DeleteKey doesn't work if there are subkeys
-		registry.DeleteKey(registry.CURRENT_USER, `SOFTWARE\Classes\Applications\Explorer.exe\Drives\`+oldMount[:1]+`\DefaultIcon`)
-		registry.DeleteKey(registry.CURRENT_USER, `SOFTWARE\Classes\Applications\Explorer.exe\Drives\`+oldMount[:1]+`\DefaultLabel`)
-		registry.DeleteKey(registry.CURRENT_USER, `SOFTWARE\Classes\Applications\Explorer.exe\Drives\`+oldMount[:1])
-		notifyShell(oldMount)
-	}
-	if newMount == "" {
-		return nil
-	}
-	k, _, err := registry.CreateKey(registry.CURRENT_USER, `SOFTWARE\Classes\Applications\Explorer.exe\Drives\`+newMount[:1]+`\DefaultIcon`, registry.SET_VALUE|registry.CREATE_SUB_KEY|registry.WRITE)
-	defer k.Close()
-	if err != nil {
-		return err
-	}
-	keybaseExe, err := osext.Executable()
-	if err != nil {
-		return err
-	}
-	// Use the second icon bound into keybase.exe - hence the 1
-	err = k.SetStringValue("", keybaseExe+",1")
-	if err != nil {
-		return err
-	}
-
-	// Also give a nice label
-	k2, _, err := registry.CreateKey(registry.CURRENT_USER, `SOFTWARE\Classes\Applications\Explorer.exe\Drives\`+newMount[:1]+`\DefaultLabel`, registry.SET_VALUE|registry.CREATE_SUB_KEY|registry.WRITE)
-	defer k2.Close()
-	err = k2.SetStringValue("", "Keybase")
-	notifyShell(newMount)
-	return err
 }
